Move TestParseBlockSizeMatchesBytes into external_test.go
The test covers ParsedBlock's Size and Bytes, which live in external.go, so it
belongs beside them rather than in the instance tests. The body is unchanged.
